tools refactor pt 2
